analisis/requirements.txt
flask
flask_sqlalchemy
flask_cors
pandas
numpy
scipy
scikit-learn
mysql-connector-python
python-dotenv
